refactor(core): Move major top level functions to their own files
