+++ core/sile.lua
@@ -10,9 +10,6 @@ require("core.globals")
 -- Lua loader modifications to support more resource types and module layouts
 require("core.loaders")
 
--- Reserve scope placeholder for profiler (developer tooling)
-local ProFi
-
 -- Placeholder for SILE internals table
 SILE = {}
 
@@ -181,19 +178,6 @@ SILE.papersize = require("core.papersize")
 -- they run core SILE functions on load instead of waiting to be called (or
 -- depend on others that do).
 
-local function runEvals (evals, arg)
-   for _, snippet in ipairs(evals) do
-      local pId = SILE.traceStack:pushText(snippet)
-      local status, func = pcall(load, snippet)
-      if status and type(func) == "function" then
-         func()
-      else
-         SU.error(("Error parsing code provided in --%s snippet: %s"):format(arg, snippet))
-      end
-      SILE.traceStack:pop(pId)
-   end
-end
-
 --- Core functions
 -- @section functions
 
@@ -208,89 +192,7 @@ end
 -- 6. Runs any code snippents passed with `--eval`.
 --
 -- Does not move on to processing input document(s).
-function SILE.init ()
-   if SILE.backend then
-      SU.deprecated("SILE.backend", "SILE.input.backend", "0.15.7", "0.17.0")
-      SILE.input.backend = SILE.backend
-   end
-   for _, tree in ipairs(SILE.input.luarocksTrees) do
-      _G["extendSilePathRocks"](tree)
-   end
-   if not SILE.input.backend then
-      SILE.input.backend = "libtexpdf"
-   end
-   if SILE.input.backend == "libtexpdf" then
-      SILE.shaper = SILE.shapers.harfbuzz()
-      SILE.outputter = SILE.outputters.libtexpdf()
-   elseif SILE.input.backend == "cairo" then
-      SILE.shaper = SILE.shapers.pango()
-      SILE.outputter = SILE.outputters.cairo()
-   elseif SILE.input.backend == "debug" then
-      SILE.shaper = SILE.shapers.harfbuzz()
-      SILE.outputter = SILE.outputters.debug()
-   elseif SILE.input.backend == "text" then
-      SILE.shaper = SILE.shapers.harfbuzz()
-      SILE.outputter = SILE.outputters.text()
-   elseif SILE.input.backend == "dummy" then
-      SILE.shaper = SILE.shapers.harfbuzz()
-      SILE.outputter = SILE.outputters.dummy()
-   end
-   SILE.pagebuilder = SILE.pagebuilders.default()
-   io.stdout:setvbuf("no")
-   if SU.debugging("profile") then
-      ProFi = require("ProFi")
-      ProFi:start()
-   end
-   if SILE.makeDeps then
-      SILE.makeDeps:add(_G.executablePath)
-   end
-   runEvals(SILE.input.evaluates, "evaluate")
-end
-
-local function suggest_luarocks (module)
-   local guessed_module_name = module:gsub(".*%.", "") .. ".sile"
-   return ([[
-
-      If the expected module is a 3rd party extension you may need to install
-      it using LuaRocks. The details of how to do this are highly dependent on
-      your system and preferred installation method, but as an example
-      installing a 3rd party SILE module to a project-local tree where might
-      look like this:
-
-        luarocks --lua-version %s --tree lua_modules install %s
-
-      This will install the LuaRock(s) to your project. Note this takes
-      advantage of the fact that SILE checks for modules in the path
-      'lua_modules' relative to the current input file by default. SILE also
-      automatically checks the default system Lua path by default, so using
-      `--global` will also work.
-
-      In the event you use a different path to the LuaRocks tree, you must also
-      set an environment variable to teach SILE about how to find the tree
-      *before* it runs. This can be aided by asking LuaRocks to come up with a
-      path and evaling the result in the shell before running SILE. This only
-      needs to be done once in each shell, (obviously substituting 'path' for
-      your actual path):
-
-        eval $(luarocks --lua-version %s --tree path)
-
-      Thereafter running `sile` as normal in the same shell should work as
-      expected. This code can be used in your shell's initialization script
-      to avoid having to do it manually in each new shell. This is true for
-      user home directory installations using `--local` or any specific values
-      for `--tree` other than 'lua_modules'.
-
-      As an anternative to setting up environment variables when using a
-      non-default tree location, you can use the `--luarocks-tree` option to
-      add path(s) at runtime. This is simpler to type, but must be used on each
-      and every invocation. The value for tree should be the same as used when
-      installing the LuaRock(s), or an appropriate full path to the location
-      used by `--local` (generally "$HOME/.luarocks"):
-
-        sile --luarocks-tree path %s
-
-    ]]):format(SILE.lua_version, guessed_module_name, SILE.lua_version, pl.stringx.join(" ", _G.arg or {}))
-end
+SILE.init = require("core.init").init
 
 --- Multi-purpose loader to load and initialize modules.
 -- This is used to load and initialize core parts of SILE and also 3rd party modules.
@@ -299,195 +201,18 @@ end
 --   a module that has already been loaded.
 -- @tparam[opt] table options Startup options as key/value pairs passed to the module when initialized.
 -- @tparam[opt=false] boolean reload whether or not to reload a module that has been loaded and initialized before.
-function SILE.use (module, options, reload)
-   local status, pack
-   if type(module) == "string" then
-      if module:match("/") then
-         SU.warn(([[
-            Module names should not include platform-specific path separators
-
-            Using slashes like '/' or '\' in a module name looks like a path segment. This
-            may appear to work in some cases, but breaks cross platform compatibility.
-            Even on the platform with the matching separator, this can lead to packages
-            getting loaded more than once because Lua will cache one each of the different
-            formats. Please use '.' separators which are automatically translated to the
-            correct platform one. For example a correct use statement would be:
-
-              \use[module=%s] instead of \use[module=%s].
-         ]]):format(module:gsub("/", "."), module))
-      end
-      status, pack = pcall(require, module)
-      if not status then
-         SU.error(
-            ("Unable to use '%s':\n%s%s"):format(
-               module,
-               SILE.traceback and ("    Lua " .. pack) or "",
-               suggest_luarocks(module)
-            )
-         )
-      end
-   elseif type(module) == "table" then
-      pack = module
-   end
-   local name = pack._name
-   local class = SILE.documentState.documentClass
-   if not pack.type then
-      SU.error("Modules must declare their type")
-   elseif pack.type == "class" then
-      SILE.classes[name] = pack
-      if class then
-         SU.error("Cannot load a class after one is already instantiated")
-      end
-      SILE.scratch.class_from_uses = pack
-   elseif pack.type == "inputter" then
-      SILE.inputters[name] = pack
-      SILE.inputter = pack(options)
-   elseif pack.type == "outputter" then
-      SILE.outputters[name] = pack
-      SILE.outputter = pack(options)
-   elseif pack.type == "shaper" then
-      SILE.shapers[name] = pack
-      SILE.shaper = pack(options)
-   elseif pack.type == "typesetter" then
-      SILE.typesetters[name] = pack
-      SILE.typesetter = pack(options)
-   elseif pack.type == "pagebuilder" then
-      SILE.pagebuilders[name] = pack
-      SILE.pagebuilder = pack(options)
-   elseif pack.type == "package" then
-      SILE.packages[pack._name] = pack
-      if class then
-         class:loadPackage(pack, options, reload)
-      else
-         table.insert(SILE.input.preambles, { pack = pack, options = options })
-      end
-   end
-end
+SILE.use = require("core.use")
 
 -- --- Content loader like Lua's `require()` but with special path handling for loading SILE resource files.
 -- -- Used for example by commands that load data via a `src=file.name` option.
 -- -- @tparam string dependency Lua spec
-function SILE.require (dependency, pathprefix, deprecation_ack)
-   if pathprefix and not deprecation_ack then
-      local notice = string.format(
-         [[
-  Please don't use the path prefix mechanism; it was intended to provide
-  alternate paths to override core components but never worked well and is
-  causing portability problems. Just use Lua idiomatic module loading:
-      SILE.require("%s", "%s") → SILE.require("%s.%s")]],
-         dependency,
-         pathprefix,
-         pathprefix,
-         dependency
-      )
-      SU.deprecated("SILE.require", "SILE.require", "0.13.0", nil, notice)
-   end
-   dependency = dependency:gsub(".lua$", "")
-   local status, lib
-   if pathprefix then
-      -- Note this is not a *path*, it is a module identifier:
-      -- https://github.com/sile-typesetter/sile/issues/1861
-      status, lib = pcall(require, pl.stringx.join(".", { pathprefix, dependency }))
-   end
-   if not status then
-      local prefixederror = lib
-      status, lib = pcall(require, dependency)
-      if not status then
-         SU.error(
-            ("Unable to find module '%s'%s"):format(
-               dependency,
-               SILE.traceback and ((pathprefix and "\n  " .. prefixederror or "") .. "\n  " .. lib) or ""
-            )
-         )
-      end
-   end
-   local class = SILE.documentState.documentClass
-   if not class and not deprecation_ack then
-      SU.warn(string.format(
-         [[
-            SILE.require() is only supported in documents, packages, or class init
-
-            It will not function fully before the class is instantiated. Please just use
-            the Lua require() function directly:
-
-              SILE.require("%s") → require("%s")
-         ]],
-         dependency,
-         dependency
-      ))
-   end
-   if type(lib) == "table" and class then
-      if lib.type == "package" then
-         lib(class)
-      else
-         class:initPackage(lib)
-      end
-   end
-   return lib
-end
+SILE.require = require("core.require").require
 
 --- Process content.
 -- This is the main 'action' SILE does. Once input files are parsed into an abstract syntax tree, then we recursively
 -- iterate through the tree handling each item in the order encountered.
 -- @tparam table ast SILE content in abstract syntax tree format (a table of strings, functions, or more AST trees).
-function SILE.process (ast)
-   if not ast then
-      return
-   end
-   if SU.debugging("ast") then
-      SU.debugAST(ast, 0)
-   end
-   if type(ast) == "function" then
-      return ast()
-   end
-   for _, content in ipairs(ast) do
-      if type(content) == "string" then
-         SILE.typesetter:typeset(content)
-      elseif type(content) == "function" then
-         content()
-      elseif SILE.Commands[content.command] then
-         SILE.call(content.command, content.options, content)
-      elseif not content.command and not content.id then
-         local pId = SILE.traceStack:pushContent(content, "content")
-         SILE.process(content)
-         SILE.traceStack:pop(pId)
-      elseif type(content) ~= "nil" then
-         local pId = SILE.traceStack:pushContent(content)
-         SU.error("Unknown command " .. (tostring(content.command or content.id)))
-         SILE.traceStack:pop(pId)
-      end
-   end
-end
-
-local preloadedinputters = { "xml", "lua", "sil" }
-
-local function detectFormat (doc, filename)
-   -- Preload default reader types so content detection has something to work with
-   if #SILE.inputters == 0 then
-      for _, format in ipairs(preloadedinputters) do
-         local _ = SILE.inputters[format]
-      end
-   end
-   local contentDetectionOrder = {}
-   for _, inputter in pairs(SILE.inputters) do
-      if inputter.order then
-         table.insert(contentDetectionOrder, inputter)
-      end
-   end
-   table.sort(contentDetectionOrder, function (a, b)
-      return a.order < b.order
-   end)
-   local initialround = filename and 1 or 2
-   for round = initialround, 3 do
-      for _, inputter in ipairs(contentDetectionOrder) do
-         SU.debug("inputter", "Running content type detection round", round, "with", inputter._name)
-         if inputter.appropriate(round, filename, doc) then
-            return inputter._name
-         end
-      end
-   end
-   SU.error(("Unable to pick inputter to process input from '%s'"):format(filename))
-end
+SILE.process = require("core.process").process
 
 --- Process an input string.
 -- First converts the string to an AST, then runs `process` on it.
@@ -495,42 +220,7 @@ end
 -- @tparam[opt] nil|string format The name of the formatter. If nil, defaults to using each intputter's auto detection.
 -- @tparam[opt] nil|string filename Pseudo filename to identify the content with, useful for error messages stack traces.
 -- @tparam[opt] nil|table options Options to pass to the inputter instance when instantiated.
-function SILE.processString (doc, format, filename, options)
-   local cpf
-   if not filename then
-      cpf = SILE.currentlyProcessingFile
-      local caller = debug.getinfo(2, "Sl")
-      SILE.currentlyProcessingFile = caller.short_src .. ":" .. caller.currentline
-   end
-   -- In the event we're processing the master file *and* the user gave us
-   -- a specific inputter to use, use it at the exclusion of all content type
-   -- detection
-   local inputter
-   if
-      filename
-      and pl.path.normcase(pl.path.normpath(filename)) == pl.path.normcase(SILE.input.filenames[1])
-      and SILE.inputter
-   then
-      inputter = SILE.inputter
-   else
-      format = format or detectFormat(doc, filename)
-      if not SILE.quiet then
-         io.stderr:write(("<%s> as %s\n"):format(SILE.currentlyProcessingFile, format))
-      end
-      inputter = SILE.inputters[format](options)
-      -- If we did content detection *and* this is the master file, save the
-      -- inputter for posterity and postambles
-      if filename and pl.path.normcase(filename) == pl.path.normcase(SILE.input.filenames[1]:gsub("^-$", "STDIN")) then
-         SILE.inputter = inputter
-      end
-   end
-   local pId = SILE.traceStack:pushDocument(SILE.currentlyProcessingFile, doc)
-   inputter:process(doc)
-   SILE.traceStack:pop(pId)
-   if cpf then
-      SILE.currentlyProcessingFile = cpf
-   end
-end
+SILE.processString = require("core.process").processString
 
 --- Process an input file
 -- Opens a file, converts the contents to an AST, then runs `process` on it.
@@ -538,103 +228,17 @@ end
 -- @tparam string filename Path of file to open string to be converted to SILE content.
 -- @tparam[opt] nil|string format The name of the formatter. If nil, defaults to using each intputter's auto detection.
 -- @tparam[opt] nil|table options Options to pass to the inputter instance when instantiated.
-function SILE.processFile (filename, format, options)
-   local lfs = require("lfs")
-   local doc
-   if filename == "-" then
-      filename = "STDIN"
-      doc = io.stdin:read("*a")
-   else
-      -- Turn slashes around in the event we get passed a path from a Windows shell
-      filename = filename:gsub("\\", "/")
-      if not SILE.masterFilename then
-         SILE.masterFilename = pl.path.splitext(pl.path.normpath(filename))
-      end
-      if SILE.input.filenames[1] and not SILE.masterDir then
-         SILE.masterDir = pl.path.dirname(SILE.input.filenames[1])
-      end
-      if SILE.masterDir and SILE.masterDir:len() >= 1 then
-         _G.extendSilePath(SILE.masterDir)
-         _G.extendSilePathRocks(SILE.masterDir .. "/lua_modules")
-      end
-      filename = SILE.resolveFile(filename) or SU.error("Could not find file")
-      local mode = lfs.attributes(filename).mode
-      if mode ~= "file" and mode ~= "named pipe" then
-         SU.error(filename .. " isn't a file or named pipe, it's a " .. mode .. "!")
-      end
-      if SILE.makeDeps then
-         SILE.makeDeps:add(filename)
-      end
-      local file, err = io.open(filename)
-      if not file then
-         print("Could not open " .. filename .. ": " .. err)
-         return
-      end
-      doc = file:read("*a")
-   end
-   local cpf = SILE.currentlyProcessingFile
-   SILE.currentlyProcessingFile = filename
-   local pId = SILE.traceStack:pushDocument(filename, doc)
-   local ret = SILE.processString(doc, format, filename, options)
-   SILE.traceStack:pop(pId)
-   SILE.currentlyProcessingFile = cpf
-   return ret
-end
+SILE.processFile = require("core.process").processFile
 
 -- TODO: this probably needs deprecating, moved here just to get out of the way so
 -- typesetters classing works as expected
-function SILE.typesetNaturally (frame, func)
-   local saveTypesetter = SILE.typesetter
-   if SILE.typesetter.frame then
-      SILE.typesetter.frame:leave(SILE.typesetter)
-   end
-   SILE.typesetter = SILE.typesetters.default(frame)
-   SILE.settings:temporarily(func)
-   SILE.typesetter:leaveHmode()
-   SILE.typesetter:chuck()
-   SILE.typesetter.frame:leave(SILE.typesetter)
-   SILE.typesetter = saveTypesetter
-   if SILE.typesetter.frame then
-      SILE.typesetter.frame:enter(SILE.typesetter)
-   end
-end
+SILE.typesetNaturally = require("core.misc").typesetNaturally
 
 --- Resolve relative file paths to identify absolute resources locations.
 -- Makes it possible to load resources from relative paths, relative to a document or project or SILE itself.
 -- @tparam string filename Name of file to find using the same order of precedence logic in `require()`.
 -- @tparam[opt] nil|string pathprefix Optional prefix in which to look for if the file isn't found otherwise.
-function SILE.resolveFile (filename, pathprefix)
-   local candidates = {}
-   -- Start with the raw file name as given prefixed with a path if requested
-   if pathprefix then
-      candidates[#candidates + 1] = pl.path.join(pathprefix, "?")
-   end
-   -- Also check the raw file name without a path
-   candidates[#candidates + 1] = "?"
-   -- Iterate through the directory of the master file, the SILE_PATH variable, and the current directory
-   -- Check for prefixed paths first, then the plain path in that fails
-   if SILE.masterDir then
-      for path in SU.gtoke(SILE.masterDir .. ";" .. tostring(os.getenv("SILE_PATH")), ";") do
-         if path.string and path.string ~= "nil" then
-            if pathprefix then
-               candidates[#candidates + 1] = pl.path.join(path.string, pathprefix, "?")
-            end
-            candidates[#candidates + 1] = pl.path.join(path.string, "?")
-         end
-      end
-   end
-   -- Return the first candidate that exists, also checking the .sil suffix
-   local path = table.concat(candidates, ";")
-   local resolved, err = package.searchpath(filename, path, "/")
-   if resolved then
-      if SILE.makeDeps then
-         SILE.makeDeps:add(resolved)
-      end
-   elseif SU.debugging("paths") then
-      SU.debug("paths", ("Unable to find file '%s': %s"):format(filename, err))
-   end
-   return resolved
-end
+SILE.resolveFile = require("core.require").resolveFile
 
 --- Execute a registered SILE command.
 -- Uses a function previously registered by any modules explicitly loaded by the user at runtime via `--use`, the SILE
@@ -642,23 +246,7 @@ end
 -- @tparam string command Command name.
 -- @tparam[opt={}] nil|table options Options to pass to the command.
 -- @tparam[opt] nil|table content Any valid AST node to be processed by the command.
-function SILE.call (command, options, content)
-   options = options or {}
-   content = content or {}
-   if SILE.traceback and type(content) == "table" and not content.lno then
-      -- This call is from code (no content.lno) and we want to spend the time
-      -- to determine everything we need about the caller
-      local caller = debug.getinfo(2, "Sl")
-      content.file, content.lno = caller.short_src, caller.currentline
-   end
-   local pId = SILE.traceStack:pushCommand(command, content, options)
-   if not SILE.Commands[command] then
-      SU.error("Unknown command " .. command)
-   end
-   local result = SILE.Commands[command](options, content)
-   SILE.traceStack:pop(pId)
-   return result
-end
+SILE.call = require("core.misc").call
 
 --- (Deprecated) Register a function as a SILE command.
 -- Takes any Lua function and registers it for use as a SILE command (which will in turn be used to process any content
@@ -673,55 +261,19 @@ end
 -- messages. Usually auto-detected.
 -- @see SILE.classes:registerCommand
 -- @see SILE.packages:registerCommand
-function SILE.registerCommand (name, func, help, pack, cheat)
-   local class = SILE.documentState.documentClass
-   if not cheat then
-      SU.deprecated(
-         "SILE.registerCommand",
-         "class:registerCommand / package:registerCommand",
-         "0.14.0",
-         "0.16.0",
-         [[
-            Commands are being scoped to the document classes or packages they are
-            loaded into rather than using a global registry.
-         ]]
-      )
-   end
-   -- Shimming until we have all scope cheating removed from core
-   if not cheat or not class or class.type ~= "class" then
-      return SILE.classes.base.registerCommand(nil, name, func, help, pack)
-   end
-   return class:registerCommand(name, func, help, pack)
-end
+SILE.registerCommand = require("core.misc").registerCommand
 
 --- Wrap an existing command with new default options.
 -- Modifies an already registered SILE command with a new table of options to be used as default values any time it is
 -- called. Calling options still take precedence.
 -- @tparam string command Name of command to overwrite.
 -- @tparam table options Options to set as updated defaults.
-function SILE.setCommandDefaults (command, options)
-   local oldCommand = SILE.Commands[command]
-   SILE.Commands[command] = function (defaults, content)
-      for k, v in pairs(options) do
-         defaults[k] = defaults[k] or v
-      end
-      return oldCommand(defaults, content)
-   end
-end
+SILE.setCommandDefaults = require("core.misc").setCommandDefaults
 
 -- TODO: Move to new table entry handler in types.unit
-function SILE.registerUnit (unit, spec)
-   -- If a unit exists already, clear it first so we get fresh meta table entries, see #1607
-   if SILE.types.unit[unit] then
-      SILE.types.unit[unit] = nil
-   end
-   SILE.types.unit[unit] = spec
-end
+SILE.registerUnit = require("core.misc").registerUnit
 
-function SILE.paperSizeParser (size)
-   SU.deprecated("SILE.paperSizeParser", "SILE.papersize", "0.15.0", "0.16.0")
-   return SILE.papersize(size)
-end
+SILE.paperSizeParser = require("core.misc").paperSizeParser
 
 --- Finalize document processing
 -- Signals that all the `SILE.process()` calls have been made and SILE should move on to finish up the output
@@ -733,24 +285,7 @@ end
 -- 4. Stops logging dependencies and writes them to a makedepends file if requested.
 -- 5. Close out the Lua profiler if it was running.
 -- 6. Output version information if versions debug flag is set.
-function SILE.finish ()
-   SILE.documentState.documentClass:finish()
-   SILE.font.finish()
-   runEvals(SILE.input.evaluateAfters, "evaluate-after")
-   if SILE.makeDeps then
-      SILE.makeDeps:write()
-   end
-   if not SILE.quiet then
-      io.stderr:write("\n")
-   end
-   if SU.debugging("profile") then
-      ProFi:stop()
-      ProFi:writeReport(pl.path.splitext(SILE.input.filenames[1]) .. ".profile.txt")
-   end
-   if SU.debugging("versions") then
-      SILE.shaper:debugVersions()
-   end
-end
+SILE.finish = require("core.init").finish
 
 -- Internal libraries that return classes, but we have no subclasses an only ever use one instantiation of the base
 SILE.traceStack = require("core.tracestack")()
